backend/requirements.txt
Flask==1.1.1
itsdangerous==1.1.0
Flask-Mail==0.9.1
jsonpickle==1.3
pymongo==3.11.0
numpy==1.18.1
dnspython==1.16.0
gunicorn==20.0.4
flask-cors==3.0.9
Pillow==7.2.0